requirements.txt
djangorestframework
django-environ==0.9.0
psycopg2
Django
django-cors-headers